requirements.txt
# Core
torch>=2.0.0
transformers>=4.45.0  # Required for Qwen2-VL
accelerate>=0.27.0

# Vision models
timm>=0.9.0
einops>=0.7.0

# Claude API
anthropic>=0.25.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.9.0

# Data handling
pyyaml>=6.0
pandas>=2.0.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0

# Optional: Flash attention for faster inference
# pip install flash-attn --no-build-isolation

# Qwen3-VL (requires transformers from source):
#   pip install git+https://github.com/huggingface/transformers

# DeepSeek-VL2 (install separately):
#   git clone https://github.com/deepseek-ai/DeepSeek-VL2
#   cd DeepSeek-VL2 && pip install -e .
